Refactor sendIpcMessage into own file
